Add Color-Coded Feedback to Installation Functions
In this commit, we have updated the PowerShell functions responsible for installing various software packages (Google Chrome, TeamViewer, Adobe Reader, Java, FortiClient VPN, 7-Zip, and Notepad++) to provide color-coded feedback to the user. The changes are as follows:

1. Enhanced user feedback by adding color-coded messages to indicate the success or failure of the installation process. 
2. In the event of a failure during any phase of the installation (download, install, or cleanup), a red error message is displayed to the user.
3. Upon successful installation, a green message is displayed to indicate success.

These changes aim to improve the user experience by providing clear and immediate feedback on the outcome of the installation process.

Files Modified:
- Install-GoogleChrome.ps1
- Install-TeamViewer.ps1
- Install-AdobeReader.ps1
- Install-Java.ps1
- Install-FortiClientVPN.ps1
- Install-7Zip.ps1
- Install-NotepadPlusPlus.ps1

We believe that these updates will enhance the usability and user-friendliness of the installation scripts.